Baked Lighting + Reflections
The lighting is baked this time round (It wasn't previously), with
reflections. Some areas still not as desirable as the one in my own
project folder... I had forgotten to mark some props as lightmap static
too so it's not complete.

Otherwise, the things inside should be updated with my progress so far.