Mainly Tweaks
Will be finalizing Interior Gameflow during tomorrow (Tie up loose ends
and voice over) and prefab it for export.
Exterior will also be done, but during the weekend.

Scene Changes:
- Tweaked the flow of interior a little due to feedback (Teach players
run a lot earlier to save time [Since the player is late for the meeting
it makes sense to run])
- Blocked off the Exterior Gameflow for second pathing cause no assets
for it yet (Will be super blank)
- Had to fix the unity crash issue and save the scene (Added range to
the AI values)

Prefab Changes:
- Added Material to Weapon (yay no more pink weaps!)
- Fixed issue where hit UI was not rendering correctly (i hope)
- Added second set of footsteps to Hana Prefab (Concrete footsteps for
atrium and exterior)

Folder Changes:
-Removed obsolete test scenes Exterior and Interior