Animations and Setup
Added Run animations.

- Removed the previous walk animations and swapped them out for a
cleaner one without the model imported. (This applies to Idle and Run
animations as well)
- I also changed the animations inside the GoldHanaAC's locomotion
state. Will continue to add the other animations over the weekend

Setup the final scene in which we would be using for game build. for the
interior. Will do the same for exterior once interior is complete. Will
work on interior event manager as well